Spreadsheet "list" of CMO operations.
Spreadsheet that contains "list" of CMO operations.

Not exactly a list - in the sense that an actual operation may be a row of the table, expanded or instantiated for each of the alternatives indicated by some of the columns.

Also, some "abstract" operations may be mapped to the same implementation.  E.g. in non-coherent I/O, the user does not care whether you writeback Dirty and leave Clean or Invalid behind.  

But I am finally fairly comfortable.

No stashing - no specification of other processor caches as targets, and mostly not as sources (except implicit in backwards invalidations).